Renamed folder to frontend
